Move requestHandler.py to the system directory
