+++ agb/utility.py
@@ -14,7 +14,7 @@
 #    You should have received a copy of the GNU General Public License
 #    along with AlphaGameBot.  If not, see <https://www.gnu.org/licenses/>.
 
-import agb.requestHandler
+import agb.system.requestHandler
 import discord
 from discord.ext import commands
 import logging
@@ -126,7 +126,7 @@ class UtilityCog(agb.cogwheel.Cogwheel):
         # define word
         endpoint = agb.cogwheel.getAPIEndpoint('dictionary', "GET_WORD_DEFINITION").format(word.lower())
 
-        request = agb.requestHandler.handler.get(endpoint)
+        request = agb.system.requestHandler.handler.get(endpoint)
         response = json.loads(request.text)
 
         if request.status_code == 404: # word not found
